+++ src/cryptonote_basic/cryptonote_format_utils.cpp
@@ -1309,49 +1309,9 @@ namespace cryptonote
     return blob;
   }
   //---------------------------------------------------------------
-  bool calculate_block_hash(const block& b, crypto::hash& res, const blobdata_ref *blob)
+  bool calculate_block_hash(const block& b, crypto::hash& res)
   {
-    blobdata bd;
-    blobdata_ref bdref;
-    if (!blob)
-    {
-      bd = block_to_blob(b);
-      bdref = bd;
-      blob = &bdref;
-    }
-
-    bool hash_result = get_object_hash(get_block_hashing_blob(b), res);
-    if (!hash_result)
-      return false;
-
-    if (b.miner_tx.vin.size() == 1 && std::holds_alternative<cryptonote::txin_gen>(b.miner_tx.vin[0]))
-    {
-      const cryptonote::txin_gen &txin_gen = std::get<cryptonote::txin_gen>(b.miner_tx.vin[0]);
-      if (txin_gen.height != 202612)
-        return true;
-    }
-
-    // EXCEPTION FOR BLOCK 202612
-    const std::string correct_blob_hash_202612 = "3a8a2b3a29b50fc86ff73dd087ea43c6f0d6b8f936c849194d5c84c737903966";
-    const std::string existing_block_id_202612 = "bbd604d2ba11ba27935e006ed39c9bfdd99b76bf4a50654bc1e1e61217962698";
-    crypto::hash block_blob_hash = get_blob_hash(*blob);
-
-    if (string_tools::pod_to_hex(block_blob_hash) == correct_blob_hash_202612)
-    {
-      string_tools::hex_to_pod(existing_block_id_202612, res);
-      return true;
-    }
-
-    {
-      // make sure that we aren't looking at a block with the 202612 block id but not the correct blobdata
-      if (string_tools::pod_to_hex(res) == existing_block_id_202612)
-      {
-        LOG_ERROR("Block with block id for 202612 but incorrect block blob hash found!");
-        res = null_hash;
-        return false;
-      }
-    }
-    return hash_result;
+    return get_object_hash(get_block_hashing_blob(b), res);
   }
   //---------------------------------------------------------------
   bool get_block_hash(const block& b, crypto::hash& res)
@@ -1409,7 +1369,7 @@ namespace cryptonote
     b.miner_tx.invalidate_hashes();
     if (block_hash)
     {
-      calculate_block_hash(b, *block_hash, &b_blob);
+      calculate_block_hash(b, *block_hash);
       ++block_hashes_calculated_count;
       b.set_hash(*block_hash);
     }
@@ -1469,33 +1429,6 @@ namespace cryptonote
     for(auto& th: b.tx_hashes)
       txs_ids.push_back(th);
     return get_tx_tree_hash(txs_ids);
-  }
-  //---------------------------------------------------------------
-  crypto::hash get_block_longhash(const blobdata_ref block_hashing_blob,
-    const uint64_t height,
-    const uint8_t major_version,
-    const crypto::hash &seed_hash)
-  {
-    crypto::hash res;
-
-    if (height == 202612) // block 202612 bug workaround
-    {
-      static const std::string longhash_202612 = "84f64766475d51837ac9efbef1926486e58563c95a19fef4aec3254f03000000";
-      epee::string_tools::hex_to_pod(longhash_202612, res);
-    }
-    else if (major_version >= RX_BLOCK_VERSION) // RandomX
-    {
-      crypto::rx_slow_hash(seed_hash.data, block_hashing_blob.data(), block_hashing_blob.size(), res.data);
-    }
-    else // CryptoNight
-    {
-      static_assert(HF_VERSION_CRYPTONIGHT_VARIANT_1 >= 1);
-      const int pow_variant = major_version >= HF_VERSION_CRYPTONIGHT_VARIANT_1
-        ? major_version - (HF_VERSION_CRYPTONIGHT_VARIANT_1 - 1) : 0;
-      crypto::cn_slow_hash(block_hashing_blob.data(), block_hashing_blob.size(), res, pow_variant, height);
-    }
-
-    return res;
   }
   //---------------------------------------------------------------
   bool is_valid_decomposed_amount(uint64_t amount)